# docs/openapi_v1.yaml
openapi: 3.1.0
info:
  title: NexusSklad API
  version: 0.1.0
  description: |
    First repository-tracked OpenAPI export for NexusSklad.
    Scope is intentionally focused on the current backend and mobile flows.
servers:
  - url: http://localhost:4000
    description: Local API
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register company and owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionRegisterResponse'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionLoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionRefreshResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/me:
    get:
      tags: [auth]
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/accept-invite:
    post:
      tags: [auth]
      summary: Accept invite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionAcceptInviteResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    UnauthorizedError:
      description: Authentication required or token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequestsError:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server-side error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UserRole:
      type: string
      enum: [OWNER, MANAGER, STAFF]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    AuthCompany:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    AuthUser:
      type: object
      required: [id, companyId, name, role, company]
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        role:
          $ref: '#/components/schemas/UserRole'
        company:
          $ref: '#/components/schemas/AuthCompany'
    AuthSession:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    RegisterRequest:
      type: object
      required: [companyName, ownerName, email, password]
      properties:
        companyName:
          type: string
        companyCity:
          type: [string, 'null']
        companyPhone:
          type: [string, 'null']
        ownerName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: [string, 'null']
        password:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AcceptInviteRequest:
      type: object
      required: [inviteToken, name, password]
      properties:
        inviteToken:
          type: string
        name:
          type: string
        phone:
          type: [string, 'null']
        password:
          type: string
    AuthSessionRegisterResponse:
      allOf:
        - $ref: '#/components/schemas/AuthSession'
        - $ref: '#/components/schemas/AuthRegisterMeta'
    AuthSessionLoginResponse:
      allOf:
        - $ref: '#/components/schemas/AuthSession'
        - $ref: '#/components/schemas/AuthLoginMeta'
    AuthSessionRefreshResponse:
      allOf:
        - $ref: '#/components/schemas/AuthSession'
        - $ref: '#/components/schemas/AuthRefreshMeta'
    AuthSessionAcceptInviteResponse:
      allOf:
        - $ref: '#/components/schemas/AuthSession'
        - $ref: '#/components/schemas/AuthAcceptInviteMeta'
    AuthRegisterMeta:
      type: object
      required: [module, action]
      properties:
        module:
          const: auth
        action:
          const: register
    AuthLoginMeta:
      type: object
      required: [module, action]
      properties:
        module:
          const: auth
        action:
          const: login
    AuthRefreshMeta:
      type: object
      required: [module, action]
      properties:
        module:
          const: auth
        action:
          const: refresh
    AuthAcceptInviteMeta:
      type: object
      required: [module, action]
      properties:
        module:
          const: auth
        action:
          const: accept-invite
    AuthMeResponse:
      type: object
      required: [user, module, action]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        module:
          const: auth
        action:
          const: me
...